REST API Done

+++ backend/controllers/goalControllers.js
@@ -1,20 +1,21 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
-
+const User = require("../models/userModel");
 //@desc get goals
 //@route api/goals
 //@access  Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 const postGoals = asyncHandler(async (req, res) => {
   if (!req.body.text) {
     res.status(400);
-    throw new Error("Please Add Text field");
+    throw new Error("Please Add text field");
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.status(200).json(goal);
@@ -25,10 +26,25 @@ const putGoals = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal Not Found");
   }
-  const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-  });
-  res.status(200).json(updatedGoal);
+  //check user exits or not
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(400);
+    throw new Error("User Not Exits");
+  }
+
+  //check it is the right user to modify data
+  const temp = goal.user.toString();
+  //string number compariosn occuring ifx this
+  if (temp == req.user.id) {
+    const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
+      new: true,
+    });
+    res.status(200).json(updatedGoal);
+  } else {
+    res.status(401);
+    throw new Error("Access Denied");
+  }
 });
 const deleteGoals = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(req.params.id);
@@ -36,12 +52,21 @@ const deleteGoals = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal Not Found");
   }
-  const deleteGoal = await Goal.findByIdAndDelete(req.params.id);
-  if (deleteGoal) {
-    res.status(200).json({id:req.params.id});
-  } else {
+  //check user exits or not
+  const user = await User.findById(req.user.id);
+  if (!user) {
     res.status(400);
-    throw new Error("Internal Error");
+    throw new Error("User Not Exits");
+  }
+
+  //check it is the right user to delete data
+  const temp = goal.user.toString();
+  //string number compariosn occuring ifx this
+  if (temp == req.user.id) {
+    await Goal.findByIdAndDelete(req.params.id);
+  } else {
+    res.status(401);
+    throw new Error("Access Denied");
   }
 });
